# stroom-app/src/main/resources/ui/noauth/swagger/stroom.yaml
openapi: 3.0.1
info:
  contact:
    name: stroom
    url: https://github.com/gchq/stroom
  description: Various APIs for interacting with Stroom and its data
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Stroom API
  version: v1/v2
servers:
- url: /api
security:
- ApiKeyAuth: []
paths:
...
  /entityEvent/v1/{nodeName}:
    put:
      operationId: fireEntityEvent
      parameters:
      - in: path
        name: memberUuid
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityEvent'
        description: entityevent
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                type: boolean
          description: default response
      summary: Sends an entity event
      tags:
      - Entity Events
...
  /processorTask/v1/abandon/{nodeName}:
    post:
      operationId: abandonProcessorTasks
      parameters:
      - in: path
        name: memberUuid
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessorTaskList'
        description: request
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                type: boolean
          description: default response
      summary: Abandon some tasks
      tags:
      - Processor Tasks
  /processorTask/v1/assign/{nodeName}:
    post:
      operationId: assignProcessorTasks
      parameters:
      - in: path
        name: memberUuid
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignTasksRequest'
        description: request
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessorTaskList'
          description: default response
      summary: Assign some tasks
      tags:
      - Processor Tasks
  /processorTask/v1/find:
    post:
      operationId: findProcessorTasks
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpressionCriteria'
        description: expressionCriteria
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPageProcessorTask'
          description: default response
      summary: Finds processors tasks
      tags:
      - Processor Tasks
  /processorTask/v1/summary:
    post:
      operationId: findProcessorTaskSummary
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpressionCriteria'
        description: expressionCriteria
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultPageProcessorTaskSummary'
          description: default response
      summary: Finds processor task summaries
      tags:
      - Processor Tasks
...
  /session/v1/list:
    get:
      operationId: listSessions
      parameters:
      - in: query
        name: memberUuid
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionListResponse'
          description: default response
      summary: Lists user sessions for a node, or all nodes in the cluster if nodeName
        is null
      tags:
      - Sessions
  /session/v1/logout:
    get:
      operationId: stroomLogout
      parameters:
      - in: query
        name: redirect_uri
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UrlResponse'
          description: default response
      summary: Logout of Stroom session
      tags:
      - Sessions
  /session/v1/noauth/validateSession:
    get:
      operationId: validateStroomSession
      parameters:
      - in: query
        name: redirect_uri
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidateSessionResponse'
          description: default response
      summary: Validate the current session, return a redirect Uri if invalid.
      tags:
      - Sessions
...
  /task/v1/find/{nodeName}:
    post:
      operationId: findTasks
      parameters:
      - in: path
        name: memberUuid
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FindTaskProgressRequest'
        description: request
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskProgressResponse'
          description: default response
      summary: Finds tasks for a node
      tags:
      - Tasks
  /task/v1/list/{nodeName}:
    get:
      operationId: listTasks
      parameters:
      - in: path
        name: memberUuid
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskProgressResponse'
          description: default response
      summary: Lists tasks for a node
      tags:
      - Tasks
  /task/v1/terminate/{nodeName}:
    post:
      operationId: terminateTasks
      parameters:
      - in: path
        name: memberUuid
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminateTaskProgressRequest'
        description: request
        required: true
      responses:
        default:
          content:
            application/json:
              schema:
                type: boolean
          description: default response
      summary: Terminates tasks for a node
      tags:
      - Tasks
  /task/v1/user/{nodeName}:
    get:
      operationId: listUserTasks
      parameters:
      - in: path
        name: memberUuid
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskProgressResponse'
          description: default response
      summary: Lists tasks for a node
      tags:
      - Tasks
...
components:
  schemas:
...
    AssignTasksRequest:
      type: object
      properties:
        count:
          type: integer
          format: int32
        nodeName:
          type: string
...
    CriteriaFieldSort:
      type: object
      properties:
        desc:
          type: boolean
        id:
          type: string
        ignoreCase:
          type: boolean
...
    DocRef:
      type: object
      description: A class for describing a unique reference to a 'document' in stroom.  A
        'document' is an entity in stroom such as a data source dictionary or pipeline.
      properties:
        name:
          type: string
          description: The name for the data source
          example: MyStatistic
        type:
          type: string
          description: The type of the 'document' that this DocRef refers to
          example: StroomStatsStore
        uuid:
          type: string
          description: The unique identifier for this 'document'
          example: 9f6184b4-bd78-48bc-b0cd-6e51a357f6a6
      required:
      - name
      - type
      - uuid
      xml:
        name: doc
...
    EntityEvent:
      type: object
      properties:
        action:
          type: string
          enum:
          - CREATE
          - UPDATE
          - DELETE
          - CLEAR_CACHE
        docRef:
          $ref: '#/components/schemas/DocRef'
...
    Expander:
      type: object
      properties:
        depth:
          type: integer
          format: int32
        expanded:
          type: boolean
        leaf:
          type: boolean
...
    ExpressionCriteria:
      type: object
      properties:
        expression:
          $ref: '#/components/schemas/ExpressionOperator'
        pageRequest:
          $ref: '#/components/schemas/PageRequest'
        sort:
          type: string
          writeOnly: true
        sortList:
          type: array
          items:
            $ref: '#/components/schemas/CriteriaFieldSort'
    ExpressionItem:
      type: object
      description: Base type for an item in an expression tree
      discriminator:
        propertyName: type
      properties:
        enabled:
          type: boolean
          description: Whether this item in the expression tree is enabled or not
          example: true
        type:
          type: string
      required:
      - type
    ExpressionOperator:
      type: object
      description: A logical addOperator term in a query expression tree
      properties:
        children:
          type: array
          items:
            $ref: '#/components/schemas/ExpressionItem'
          xml:
            wrapped: true
        enabled:
          type: boolean
          description: Whether this item in the expression tree is enabled or not
          example: true
        op:
          type: string
          description: The logical addOperator type
          enum:
          - AND
          - OR
          - NOT
      required:
      - op
...
    FindTaskCriteria:
      type: object
      properties:
        ancestorIdSet:
          type: array
          items:
            $ref: '#/components/schemas/TaskId'
          uniqueItems: true
        idSet:
          type: array
          items:
            $ref: '#/components/schemas/TaskId'
          uniqueItems: true
        sessionId:
          type: string
    FindTaskProgressCriteria:
      type: object
      properties:
        expandedTasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskProgress'
          uniqueItems: true
        nameFilter:
          type: string
        pageRequest:
          $ref: '#/components/schemas/PageRequest'
        sessionId:
          type: string
        sort:
          type: string
          writeOnly: true
        sortList:
          type: array
          items:
            $ref: '#/components/schemas/CriteriaFieldSort'
    FindTaskProgressRequest:
      type: object
      properties:
        criteria:
          $ref: '#/components/schemas/FindTaskProgressCriteria'
...
    Limits:
      type: object
      properties:
        durationMs:
          type: integer
          format: int64
        eventCount:
          type: integer
          format: int64
        streamCount:
          type: integer
          format: int64
      xml:
        name: limits
...
    PageRequest:
      type: object
      properties:
        length:
          type: integer
          format: int32
        offset:
          type: integer
          format: int32
    PageResponse:
      type: object
      description: Details of the page of results being returned.
      properties:
        exact:
          type: boolean
        length:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        total:
          type: integer
          format: int64
...
    Processor:
      type: object
      properties:
        createTimeMs:
          type: integer
          format: int64
        createUser:
          type: string
        deleted:
          type: boolean
        enabled:
          type: boolean
        id:
          type: integer
          format: int32
        pipelineName:
          type: string
        pipelineUuid:
          type: string
        taskType:
          type: string
        updateTimeMs:
          type: integer
          format: int64
        updateUser:
          type: string
        uuid:
          type: string
        version:
          type: integer
          format: int32
      xml:
        name: parameters
    ProcessorFilter:
      type: object
      properties:
        createTimeMs:
          type: integer
          format: int64
        createUser:
          type: string
        data:
          type: string
        deleted:
          type: boolean
        enabled:
          type: boolean
        id:
          type: integer
          format: int32
        maxMetaCreateTimeMs:
          type: integer
          format: int64
        minMetaCreateTimeMs:
          type: integer
          format: int64
        pipelineName:
          type: string
        pipelineUuid:
          type: string
        priority:
          type: integer
          format: int32
        processor:
          $ref: '#/components/schemas/Processor'
        processorFilterTracker:
          $ref: '#/components/schemas/ProcessorFilterTracker'
        processorUuid:
          type: string
        queryData:
          $ref: '#/components/schemas/QueryData'
        reprocess:
          type: boolean
        updateTimeMs:
          type: integer
          format: int64
        updateUser:
          type: string
        uuid:
          type: string
        version:
          type: integer
          format: int32
...
    ProcessorFilterTracker:
      type: object
      properties:
        eventCount:
          type: integer
          format: int64
        id:
          type: integer
          format: int32
        lastPollMs:
          type: integer
          format: int64
        lastPollTaskCount:
          type: integer
          format: int32
        maxMetaCreateMs:
          type: integer
          format: int64
        metaCount:
          type: integer
          format: int64
        metaCreateMs:
          type: integer
          format: int64
        minEventId:
          type: integer
          format: int64
        minMetaCreateMs:
          type: integer
          format: int64
        minMetaId:
          type: integer
          format: int64
        status:
          type: string
        version:
          type: integer
          format: int32
...
    ProcessorTask:
      type: object
      properties:
        createTimeMs:
          type: integer
          format: int64
        data:
          type: string
        endTimeMs:
          type: integer
          format: int64
        feedName:
          type: string
        id:
          type: integer
          format: int64
        metaId:
          type: integer
          format: int64
        nodeName:
          type: string
        processorFilter:
          $ref: '#/components/schemas/ProcessorFilter'
        startTimeMs:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - UNPROCESSED
          - ASSIGNED
          - PROCESSING
          - COMPLETE
          - FAILED
          - DELETED
        statusTimeMs:
          type: integer
          format: int64
        version:
          type: integer
          format: int32
    ProcessorTaskList:
      type: object
      properties:
        list:
          type: array
          items:
            $ref: '#/components/schemas/ProcessorTask'
        nodeName:
          type: string
    ProcessorTaskSummary:
      type: object
      properties:
        count:
          type: integer
          format: int64
        feed:
          type: string
        pipeline:
          $ref: '#/components/schemas/DocRef'
        priority:
          type: integer
          format: int32
        status:
          type: string
          enum:
          - UNPROCESSED
          - ASSIGNED
          - PROCESSING
          - COMPLETE
          - FAILED
          - DELETED
...
    QueryData:
      type: object
      properties:
        dataSource:
          $ref: '#/components/schemas/DocRef'
        expression:
          $ref: '#/components/schemas/ExpressionOperator'
        limits:
          $ref: '#/components/schemas/Limits'
        params:
          type: string
      xml:
        name: query
...
    ResultPageProcessorTask:
      type: object
      description: A page of results.
      properties:
        pageResponse:
          $ref: '#/components/schemas/PageResponse'
        values:
          type: array
          items:
            $ref: '#/components/schemas/ProcessorTask'
    ResultPageProcessorTaskSummary:
      type: object
      description: A page of results.
      properties:
        pageResponse:
          $ref: '#/components/schemas/PageResponse'
        values:
          type: array
          items:
            $ref: '#/components/schemas/ProcessorTaskSummary'
...
    SessionDetails:
      type: object
      properties:
        createMs:
          type: integer
          format: int64
        lastAccessedAgent:
          type: string
        lastAccessedMs:
          type: integer
          format: int64
        nodeName:
          type: string
        userName:
          type: string
...
    SessionListResponse:
      type: object
      properties:
        pageResponse:
          $ref: '#/components/schemas/PageResponse'
        values:
          type: array
          items:
            $ref: '#/components/schemas/SessionDetails'
...
    TaskId:
      type: object
      properties:
        id:
          type: string
        parentId:
          $ref: '#/components/schemas/TaskId'
    TaskProgress:
      type: object
      properties:
        expander:
          $ref: '#/components/schemas/Expander'
        filterMatchState:
          type: string
          enum:
          - MATCHED
          - NOT_MATCHED
        id:
          $ref: '#/components/schemas/TaskId'
        nodeName:
          type: string
        submitTimeMs:
          type: integer
          format: int64
        taskInfo:
          type: string
        taskName:
          type: string
        threadName:
          type: string
        timeNowMs:
          type: integer
          format: int64
        userName:
          type: string
    TaskProgressResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            type: string
        pageResponse:
          $ref: '#/components/schemas/PageResponse'
        values:
          type: array
          items:
            $ref: '#/components/schemas/TaskProgress'
    TerminateTaskProgressRequest:
      type: object
      properties:
        criteria:
          $ref: '#/components/schemas/FindTaskCriteria'
...
    UrlResponse:
      type: object
      properties:
        url:
          type: string
...
    ValidateSessionResponse:
      type: object
      properties:
        redirectUri:
          type: string
        userId:
          type: string
        valid:
          type: boolean
...
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: Authorization
      type: apiKey